float issue on contrato fixed

+++ router.js
@@ -209,12 +209,12 @@ apiRouter.post(
     check('monto_minimo')
       .isNumeric()
       .withMessage('El monto minimo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true }),
     check('monto_maximo')
       .isNumeric()
       .withMessage('El monto maximo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true }),
     check('lotes.*.nro')
       .isNumeric()
@@ -228,12 +228,12 @@ apiRouter.post(
     check('lotes.*.maximo')
       .isNumeric()
       .withMessage('El monto minimo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true }),
     check('lotes.*.minimo')
       .isNumeric()
       .withMessage('El monto minimo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true })
   ],
   contratoController.apiAddContrato
@@ -311,12 +311,12 @@ apiRouter.put(
     check('lotes.*.maximo')
       .isNumeric()
       .withMessage('El monto minimo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true }),
     check('lotes.*.minimo')
       .isNumeric()
       .withMessage('El monto minimo de contrato debe ser numérico')
-      .toInt()
+      .toFloat()
       .optional({ nullable: true, checkFalsy: true })
   ],
   contratoController.apiUpdateContrato
